Also aggregate total duration + compute daily data
Instead of aggregating by counting rides only, I now also aggregate by
summing the rides duration. In this way we keep more information on the
rides. I have also added a computation of the relative variation
day-by-day, to be used alongside the one with rolling data.

I have also improved the handling of the indices, by taking into account
the fact that we can have missing data so we need to reindex against the
total days in the considered period to uniformise.